requirements.txt
absl-py==2.1.0
adamp==0.3.0
addict==2.4.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alabaster==0.7.13
albumentations==1.4.0
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
apex==0.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==24.2.0
autocommand==2.2.2
av==12.3.0
babel==2.17.0
backports.tarfile==1.2.0
batchgenerators==0.21
beautifulsoup4==4.13.3
black==24.8.0
blinker==1.8.2
bs4==0.0.2
bypy==1.8.9
cachetools==5.3.2
certifi==2025.4.26
cffi==1.16.0
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.0.0
cmake==3.20.3
colorama==0.4.6
coloredlogs==15.0.1
community==1.0.0b1
configparser==7.1.0
contextlib2==21.6.0
contourpy==1.1.1
crcmod==1.7
cryptography==42.0.8
cycler==0.12.1
D3D==1.0
dask==2023.5.0
dcm2niix==1.0.20250506
DCNv4==1.0.0.post2
decord==0.6.0
dill==0.4.0
docker-pycreds==0.4.0
docutils==0.19
einops==0.8.0
ensemble-boxes==1.0.9
et-xmlfile==1.1.0
eval_type_backport==0.2.2
filelock==3.14.0
fire==0.7.1
Flask==3.0.3
fontmeta==1.6.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.2.3
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.28.0
google-auth-oauthlib==1.0.0
gql==0.2.0
graphql-core==1.1
grpcio==1.60.1
h5py==3.11.0
hausdorff==0.2.6
huggingface-hub==0.23.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.34.0
imagesize==1.4.1
imbalanced-learn==0.12.4
imblearn==0.0
imgaug==0.4.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
inflect==7.3.1
iopath==0.1.9
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jmespath==0.10.0
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
lightgbm==4.6.0
lightning==2.1.2
lightning-utilities==0.11.7
linecache2==1.0.0
lit==15.0.7
llvmlite==0.41.1
locket==1.0.0
loguru==0.7.3
Markdown==3.5.2
markdown-it-py==3.0.0
matplotlib==3.7.3
matplotlib-fontja==1.1.0
mdurl==0.1.2
MedPy==0.5.1
mkl-service==2.4.0
ml-collections==0.1.1
mmaction2==1.2.0
mmcls==0.25.0
mmcv==1.4.0
mmcv-full==1.5.0
mmdet==2.28.1
mmengine==0.10.5
mmsegmentation==0.27.0
model-index==0.1.11
monai==0.3.0
more-itertools==10.3.0
mplfonts==0.0.9
multidict==6.1.0
multiprocess==0.70.18
MultiScaleDeformableAttention==1.0
mypy-extensions==1.0.0
natten==0.15.1+torch200cu118
nibabel==5.2.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py3==7.352.0
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.2
openxlab==0.1.1
ordered-set==4.1.0
oss2==2.17.0
packaging==24.1
pandas==2.0.3
partd==1.4.1
pathspec==0.12.1
pathtools==0.1.2
pillow==10.4.0
pip_search==0.0.13
platformdirs==4.3.6
portalocker==2.10.0
prettytable==3.10.0
promise==2.3
protobuf==5.29.4
psutil==7.0.0
ptflops==0.7.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.1
pyquaternion==0.9.9
python-dateutil==2.8.2
python-louvain==0.16
pytorch-ignite==0.5.2
pytz==2023.4
PyWavelets==1.4.1
PyYAML==6.0.2
qudida==0.0.4
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.4.2
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.13.0
segmentation-mask-overlay==0.3.0
sentry-sdk==2.27.0
setproctitle==1.3.6
shapely==2.0.6
shortuuid==1.0.13
SimpleITK==2.3.1
six==1.17.0
smmap==5.0.2
snowballstemmer==3.0.1
soupsieve==2.6
Sphinx==6.2.1
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
subprocess32==3.5.4
tabpfn==0.1.11
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
thundersvm==0.3.12
tifffile==2023.7.10
timm==1.0.9
tomli==2.0.1
toolz==0.12.1
torch==2.0.0+cu118
torch-geometric==2.6.1
torchaudio==2.0.1+cu118
torchinfo==1.8.0
torchmetrics==1.4.2
torchsummary==1.5.1
torchsummaryX==1.3.0
torchvision==0.15.1+cu118
traceback2==1.4.0
triton==2.0.0
tvdcn==0.2.3
typeguard==4.3.0
typing==3.7.4.3
typing_extensions==4.13.2
tzdata==2024.1
umap-learn==0.5.7
unittest2==1.1.0
urllib3==2.2.3
wandb==0.19.10
warmup_scheduler==0.3
watchdog==4.0.2
wcwidth==0.2.13
Werkzeug==3.0.1
xlsxwriter==3.2.5
yacs==0.1.8
yapf==0.40.1
yarl==1.11.1
zipp==3.17.0
